requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apispec==5.2.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.24.22
botocore==1.27.22
cachetools==3.1.1
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.2
deepdiff==5.8.1
dill==0.3.2
distlib==0.3.6
dnslib==0.9.19
dnspython==2.2.1
ecdsa==0.17.0
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
idna==2.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
localstack==0.14.4
localstack-client==1.35
localstack-ext==0.14.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.7
ordered-set==4.1.0
pbr==5.9.0
platformdirs==3.5.1
plux==1.3.1
protobuf==3.19.4
psutil==5.9.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==6.0
requests==2.25.1
rich==12.4.4
rsa==4.8
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
stevedore==3.5.0
tabulate==0.8.10
tailer==0.4.1
tqdm==4.65.0
urllib3==1.26.9
virtualenv==20.23.0
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
